FunJarvisProject/requirements.txt
openai
SpeechRecognition
python-dotenv
PyAudio
gTTS
pyautogui
pydub
